????? ?????????? ? int ? ????????? ?????? ? ??????? - ??? ???????? ???? ? windows ?? ????? ? ???????? ? ????? FFFFFFFFFF... - ?? ??????? ????????????

+++ question/type/preg/preg_lexer.lex.php
@@ -368,7 +368,7 @@ class qtype_preg_lexer extends JLexBase  {
                 return $this->form_res(preg_parser_yyParser::OPENBRACK, $this->form_error($text, qtype_preg_node_error::SUBTYPE_SUBPATT_NAME_EXPECTED, $pos, $pos + $length - 1, $text));
             } else {
                 $error = null;
-                $num = (int)$this->map_subpattern($name, $error);
+                $num = $this->map_subpattern($name, $error);
                 if ($error !== null) {
                     return $this->form_res(preg_parser_yyParser::OPENBRACK, $error);
                 }
@@ -631,7 +631,7 @@ class qtype_preg_lexer extends JLexBase  {
     protected function map_subpattern($name, &$error) {
         if (!array_key_exists($name, $this->subpatternmap)) {   // This subpattern does not exists.
             $num = ++$this->lastsubpatt;
-            $this->subpatternmap[$name] = (int)$num;
+            $this->subpatternmap[$name] = $num;
         } else {                                                // Subpatterns with same names should have same numbers.
             $num = $this->subpatternmap[$name];
             $this->lastsubpatt++;
@@ -640,7 +640,7 @@ class qtype_preg_lexer extends JLexBase  {
             }
         }
         $this->maxsubpatt = max($this->maxsubpatt, $this->lastsubpatt);
-        return $num;
+        return (int)$num;
     }
     /**
      * Calculates the character for a \cx sequence.
@@ -6213,7 +6213,7 @@ array(
     $this->push_opt_lvl();
     $this->lastsubpatt++;
     $this->maxsubpatt = max($this->maxsubpatt, $this->lastsubpatt);
-    return $this->form_res(preg_parser_yyParser::OPENBRACK, new qtype_preg_lexem_subpatt(qtype_preg_node_subpatt::SUBTYPE_SUBPATT, $this->yychar, $this->yychar, $this->yytext(), (int)$this->lastsubpatt));
+    return $this->form_res(preg_parser_yyParser::OPENBRACK, new qtype_preg_lexem_subpatt(qtype_preg_node_subpatt::SUBTYPE_SUBPATT, $this->yychar, $this->yychar, $this->yytext(), $this->lastsubpatt));
 }
                         case -10:
                             break;
@@ -6556,7 +6556,7 @@ array(
     $this->push_opt_lvl();
     $this->lastsubpatt++;
     $this->maxsubpatt = max($this->maxsubpatt, $this->lastsubpatt);
-    return $this->form_res(preg_parser_yyParser::OPENBRACK, new qtype_preg_lexem_subpatt(qtype_preg_node_subpatt::SUBTYPE_ONCEONLY, $this->yychar, $this->yychar + $this->yylength() - 1, $this->yytext(), (int)$this->lastsubpatt));
+    return $this->form_res(preg_parser_yyParser::OPENBRACK, new qtype_preg_lexem_subpatt(qtype_preg_node_subpatt::SUBTYPE_ONCEONLY, $this->yychar, $this->yychar + $this->yylength() - 1, $this->yytext(), $this->lastsubpatt));
 }
                         case -51:
                             break;
@@ -6830,7 +6830,7 @@ array(
     if ($num < 0) {
         $num = $this->lastsubpatt + $num + 1;
     }
-    return $this->form_backref($text, $this->yychar, $this->yylength(), (int)$num);
+    return $this->form_backref($text, $this->yychar, $this->yylength(), $num);
 }
                         case -81:
                             break;
@@ -6838,7 +6838,7 @@ array(
                             {
     $text = $this->yytext();
     $str = qtype_poasquestion_string::substr($text, 3, $this->yylength() - 4);
-    $code = (int)hexdec($str);
+    $code = hexdec($str);
     if ($code > qtype_preg_unicode::max_possible_code()) {
         return $this->form_res(preg_parser_yyParser::PARSLEAF, $this->form_error($text, qtype_preg_node_error::SUBTYPE_CHAR_CODE_TOO_BIG, $this->yychar, $this->yychar + $this->yylength() - 1, '0x' . $str));
     } else {
@@ -7079,7 +7079,7 @@ array(
                             {
     $text = $this->yytext();
     $str = qtype_poasquestion_string::substr($text, 3, $this->yylength() - 4);
-    $code = (int)hexdec($str);
+    $code = hexdec($str);
     if ($code > qtype_preg_unicode::max_possible_code()) {
         $this->charset->error[] = $this->form_error($text, qtype_preg_node_error::SUBTYPE_CHAR_CODE_TOO_BIG, $this->yychar, $this->yychar + $this->yylength() - 1, '0x' . $str);
     } else {
